+++ saleor/graphql/checkout/mutations/checkout_customer_note_update.py
@@ -11,7 +11,7 @@ from ...plugins.dataloaders import get_plugin_manager_promise
 from ..types import Checkout
 
 
-class CheckoutNoteUpdate(BaseMutation):
+class CheckoutCustomerNoteUpdate(BaseMutation):
     checkout = graphene.Field(Checkout, description="An updated checkout.")
 
     class Arguments:
@@ -19,10 +19,14 @@ class CheckoutNoteUpdate(BaseMutation):
             description="The checkout's ID." + ADDED_IN_321,
             required=False,
         )
-        note = graphene.String(required=True, description="New note content.")
+        customer_note = graphene.String(
+            required=True, description="New customer note content."
+        )
 
     class Meta:
-        description = "Updates note in the existing checkout object." + ADDED_IN_321
+        description = (
+            "Updates customer note in the existing checkout object." + ADDED_IN_321
+        )
         doc_category = DOC_CATEGORY_CHECKOUT
         error_type_class = CheckoutError
         error_type_field = "checkout_errors"
@@ -40,14 +44,14 @@ class CheckoutNoteUpdate(BaseMutation):
         info: ResolveInfo,
         /,
         *,
-        note,
+        customer_note,
         id=None,
     ):
         checkout = cls.get_node_or_error(info, id, only_type=Checkout)
 
-        checkout.note = note
+        checkout.note = customer_note
         cls.clean_instance(info, checkout)
         checkout.save(update_fields=["note", "last_change"])
         manager = get_plugin_manager_promise(info.context).get()
         cls.call_event(manager.checkout_updated, checkout)
-        return CheckoutNoteUpdate(checkout=checkout)
+        return CheckoutCustomerNoteUpdate(checkout=checkout)