+++ saleor/graphql/checkout/schema.py
@@ -26,6 +26,7 @@ from .mutations import (
     CheckoutCreateFromOrder,
     CheckoutCustomerAttach,
     CheckoutCustomerDetach,
+    CheckoutCustomerNoteUpdate,
     CheckoutDeliveryMethodUpdate,
     CheckoutEmailUpdate,
     CheckoutLanguageCodeUpdate,
@@ -33,7 +34,6 @@ from .mutations import (
     CheckoutLinesAdd,
     CheckoutLinesDelete,
     CheckoutLinesUpdate,
-    CheckoutNoteUpdate,
     CheckoutRemovePromoCode,
     CheckoutShippingAddressUpdate,
     CheckoutShippingMethodUpdate,
@@ -123,6 +123,7 @@ class CheckoutMutations(graphene.ObjectType):
     checkout_create_from_order = CheckoutCreateFromOrder.Field()
     checkout_customer_attach = CheckoutCustomerAttach.Field()
     checkout_customer_detach = CheckoutCustomerDetach.Field()
+    checkout_customer_note_update = CheckoutCustomerNoteUpdate.Field()
     checkout_email_update = CheckoutEmailUpdate.Field()
     checkout_line_delete = CheckoutLineDelete.Field(
         deprecation_reason=(
@@ -132,7 +133,6 @@ class CheckoutMutations(graphene.ObjectType):
     checkout_lines_delete = CheckoutLinesDelete.Field()
     checkout_lines_add = CheckoutLinesAdd.Field()
     checkout_lines_update = CheckoutLinesUpdate.Field()
-    checkout_note_update = CheckoutNoteUpdate.Field()
     checkout_remove_promo_code = CheckoutRemovePromoCode.Field()
     checkout_payment_create = CheckoutPaymentCreate.Field()
     checkout_shipping_address_update = CheckoutShippingAddressUpdate.Field()