+++ saleor/graphql/checkout/tests/mutations/test_checkout_customer_note_update.py
@@ -1,9 +1,9 @@
 from ....core.utils import to_global_id_or_none
 from ....tests.utils import get_graphql_content
 
-CHECKOUT_NOTE_UPDATE_MUTATION = """
-    mutation checkoutNoteUpdate($id: ID, $note: String!) {
-        checkoutNoteUpdate(id: $id, note: $note) {
+CHECKOUT_CUSTOMER_NOTE_UPDATE_MUTATION = """
+    mutation checkoutCustomerNoteUpdate($id: ID, $customerNote: String!) {
+        checkoutCustomerNoteUpdate(id: $id, customerNote: $customerNote) {
             checkout {
                 id,
                 note
@@ -22,25 +22,27 @@ CHECKOUT_NOTE_UPDATE_MUTATION = """
 """
 
 
-def test_checkout_note_update(user_api_client, checkout_with_item):
+def test_checkout_customer_note_update(user_api_client, checkout_with_item):
     # given
     checkout = checkout_with_item
     checkout.note = ""
     checkout.save(update_fields=["note"])
     previous_last_change = checkout.last_change
 
-    note = "New note value"
-    variables = {"id": to_global_id_or_none(checkout), "note": note}
+    customer_note = "New customer note value"
+    variables = {"id": to_global_id_or_none(checkout), "customerNote": customer_note}
 
     # when
-    response = user_api_client.post_graphql(CHECKOUT_NOTE_UPDATE_MUTATION, variables)
+    response = user_api_client.post_graphql(
+        CHECKOUT_CUSTOMER_NOTE_UPDATE_MUTATION, variables
+    )
 
     # then
     content = get_graphql_content(response)
-    data = content["data"]["checkoutNoteUpdate"]
+    data = content["data"]["checkoutCustomerNoteUpdate"]
     assert not data["errors"]
     checkout.refresh_from_db()
-    assert checkout.note == note
+    assert checkout.note == customer_note
     assert checkout.last_change != previous_last_change
 
 
@@ -52,15 +54,15 @@ def test_with_active_problems_flow(api_client, checkout_with_problems):
 
     variables = {
         "id": to_global_id_or_none(checkout_with_problems),
-        "note": "New note value",
+        "customerNote": "New customer note value",
     }
 
     # when
     response = api_client.post_graphql(
-        CHECKOUT_NOTE_UPDATE_MUTATION,
+        CHECKOUT_CUSTOMER_NOTE_UPDATE_MUTATION,
         variables,
     )
 
     # then
     content = get_graphql_content(response)
-    assert not content["data"]["checkoutNoteUpdate"]["errors"]
+    assert not content["data"]["checkoutCustomerNoteUpdate"]["errors"]