+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable, unreleased changes to this project will be documented in this file.
 
 ### GraphQL API
 
-- Add `CheckoutNoteUpdate` mutation - #16315 by @pitkes22
+- Add `CheckoutCustomerNoteUpdate` mutation - #16315 by @pitkes22
 
 ### Webhooks